moving those things had a lot of implications
