+++ packages/frontend/src/PortalForm/index.tsx
@@ -13,10 +13,10 @@ import DeviceHubIcon from '@material-ui/icons/DeviceHub'
 
 import { DEFAULT_PORTAL_SIZE, DEFAULT_ZONE } from '../common/data/constants'
 import useZoneListSelector from '../common/hooks/useZoneListSelector'
-import ZoneSearch from '../components/ZoneSearch'
-import { ZoneLight } from '../components/ZoneSearch/zoneSearchUtils'
+import ZoneSearch from '../ZoneSearch'
+import { ZoneLight } from '../ZoneSearch/zoneSearchUtils'
 import { RootState } from '../reducers'
-import { PortalSize, Zone } from '../types'
+import { PortalSize, Zone } from '../common/types'
 import UserSettings from '../UserSettings'
 import PortalSizeSelector from './PortalSizeSelector'
 import useAddPortal from './useAddPortal'
