+++ packages/frontend/src/PortalMap/ControlBar/HomeButton.tsx
@@ -4,7 +4,7 @@ import { IconButton } from '@material-ui/core'
 import HomeIcon from '@material-ui/icons/Home'
 
 import getHomeZone from '../../common/utils/getHomeZone'
-import { ZoneLight } from '../../components/ZoneSearch/zoneSearchUtils'
+import { ZoneLight } from '../../ZoneSearch/zoneSearchUtils'
 import styles from './styles.module.scss'
 
 export interface HomeButtonProps {
