+++ packages/frontend/src/PortalMap/index.tsx
@@ -5,10 +5,10 @@ import { useDispatch, useSelector } from 'react-redux'
 
 import useZoneListSelector from '../common/hooks/useZoneListSelector'
 import getHomeZone from '../common/utils/getHomeZone'
-import { ZoneLight } from '../components/ZoneSearch/zoneSearchUtils'
+import { ZoneLight } from '../ZoneSearch/zoneSearchUtils'
 import { RootState } from '../reducers'
 import { PortalMapActionTypes } from '../reducers/portalMapReducer'
-import { Zone } from '../types'
+import { Zone } from '../common/types'
 import ControlBar from './ControlBar'
 import { changeScore } from './cytoUtils'
 import defaultSettings from './defaultSettings'
