+++ packages/frontend/src/common/hooks/useZoneListSelector.ts
@@ -1,7 +1,7 @@
 import { useSelector } from 'react-redux'
 
 import { RootState } from '../../reducers'
-import { Zone } from '../../types'
+import { Zone } from '../types'
 
 const useZoneListSelector = (): Zone[] =>
   useSelector((state: RootState) => state.zones.list)
