+++ packages/frontend/src/common/hooks/useZoneSelector.ts
@@ -1,6 +1,6 @@
 import { useMemo } from 'react'
 
-import { Zone } from '../../types'
+import { Zone } from '../types'
 import useZoneListSelector from './useZoneListSelector'
 
 const useZoneSelector = (name: string): Zone | null => {
